+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-05-11T12:05:41.589Z",
+    "2026-05-11T13:44:33.441Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {